Fix: test method coverage 조건 삭제
